Various tweaks/fixes.
Refactored proxy information hash function.
Added a method for re-hashing proxies on the database.
Improved database migration from version 1 to version 2 (execute SQL to trigger it: update proxies.version set val = 1;)
Changed proxy insertion method to also replace proxies in the database if the hash is different - fixes a possible issue with proxies that use authentication and the credentials changed but not the ip/port.